Ajouter fichiers et securite

+++ src/Controller/ReservationController.php
@@ -12,11 +12,15 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
-#[Route('/reservation')]
+/**
+ * @Route("/reservation")
+ */
 class ReservationController extends AbstractController
 {
-    #[Route('/', name: 'app_reservation_index', methods: ['GET'])]
-    #[IsGranted('ROLE_CLIENT')]
+    /**
+     * @Route("/", name="app_reservation_index", methods={"GET"})
+     * @IsGranted("ROLE_CLIENT")
+     */
     public function index(ReservationRepository $reservationRepository): Response
     {
         $user = $this->getUser();
@@ -27,8 +31,10 @@ class ReservationController extends AbstractController
         ]);
     }
 
-    #[Route('/new', name: 'app_reservation_new', methods: ['GET', 'POST'])]
-    #[IsGranted('ROLE_CLIENT')]
+    /**
+     * @Route("/new", name="app_reservation_new", methods={"GET", "POST"})
+     * @IsGranted("ROLE_CLIENT")
+     */
     public function new(Request $request, EntityManagerInterface $em): Response
     {
         $reservation = new Reservation();
@@ -38,9 +44,8 @@ class ReservationController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            // Calcul automatique du prix total si tu veux
             $diff = $reservation->getDateDebut()->diff($reservation->getDateFin())->days;
-            $prixParJour = $reservation->getVehicule()->getPrixJournalier(); // à adapter selon ton entité
+            $prixParJour = $reservation->getVehicule()->getPrixParJour();
             $reservation->setPrixTotal($diff * $prixParJour);
 
             $em->persist($reservation);
@@ -54,14 +59,24 @@ class ReservationController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_reservation_edit', methods: ['GET', 'POST'])]
-    #[IsGranted('ROLE_CLIENT')]
+    /**
+     * @Route("/{id}/edit", name="app_reservation_edit", methods={"GET", "POST"})
+     * @IsGranted("ROLE_CLIENT")
+     */
     public function edit(Request $request, Reservation $reservation, EntityManagerInterface $em): Response
     {
+        if ($reservation->getUtilisateur() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('You do not own this reservation.');
+        }
+
         $form = $this->createForm(ReservationType::class, $reservation);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $diff = $reservation->getDateDebut()->diff($reservation->getDateFin())->days;
+            $prixParJour = $reservation->getVehicule()->getPrixParJour();
+            $reservation->setPrixTotal($diff * $prixParJour);
+
             $em->flush();
 
             return $this->redirectToRoute('app_reservation_index');
@@ -73,11 +88,17 @@ class ReservationController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_reservation_delete', methods: ['POST'])]
-    #[IsGranted('ROLE_CLIENT')]
+    /**
+     * @Route("/{id}", name="app_reservation_delete", methods={"POST"})
+     * @IsGranted("ROLE_CLIENT")
+     */
     public function delete(Request $request, Reservation $reservation, EntityManagerInterface $em): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$reservation->getId(), $request->request->get('_token'))) {
+        if ($reservation->getUtilisateur() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('You do not own this reservation.');
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $reservation->getId(), $request->request->get('_token'))) {
             $em->remove($reservation);
             $em->flush();
         }
@@ -85,3 +106,4 @@ class ReservationController extends AbstractController
         return $this->redirectToRoute('app_reservation_index');
     }
 }
+
